chỉnh lại homeUI, các trang liên quan home về cùng 1 style(myorder,cart,productpage...) , add thêm điều kiện không trùng code khi tạo sản phẩm, chỉnh lại creatproduct admin UI,

+++ backend/routes/UserRoute.js
@@ -48,13 +48,16 @@ router.get("/:id", auth, isAdmin, async (req, res) => {
 
 router.delete("/:id", auth, isAdmin, async (req, res) => {
   try {
+    if (req.user._id.toString() === req.params.id) {
+      return res.status(400).send("Không thể tự xóa tài khoản admin!");
+    }
     const deletedUser = await User.findByIdAndDelete(req.params.id);
 
     if (!deletedUser) {
-      return res.status(404).send("User not found");
+      return res.status(404).send("Không tìm thấy người dùng!");
     }
 
-    res.status(200).send("User deleted successfully");
+    res.status(200).send("Người dùng đã được xóa thành công!");
   } catch (err) {
     console.log(err);
